chore: Update crate dependencies

+++ src/event_graph/util.rs
@@ -103,7 +103,7 @@ pub fn seconds_until_next_rotation(next_rotation: u64) -> u64 {
     // Store `now` in a variable in order to avoid a TOCTOU error.
     // There may be a drift of one second between this panic check and
     // the return value if we get unlucky.
-    let now = UNIX_EPOCH.elapsed().unwrap().as_secs() as u64;
+    let now = UNIX_EPOCH.elapsed().unwrap().as_secs();
     if next_rotation < now {
         panic!("Next rotation timestamp is in the past");
     }
